:recycle: : 작가 호출 로직 리팩토링 - 기존 data 레이어 구조 일관성 부족으로 리팩토링
