+++ app/src/main/kotlin/com/hm/picplz/data/api/PhotographerApi.kt
@@ -1,0 +1,5 @@
+package com.hm.picplz.data.api
+
+class PhotographerApi {
+    // Todo: 작가 호출 api 정의
+}
