+++ app/src/main/kotlin/com/hm/picplz/data/source/PhotographerSource.kt
@@ -1,7 +1,7 @@
 package com.hm.picplz.data.source
 
 import com.hm.picplz.data.model.PhotographerListResponse
-import com.hm.picplz.data.service.PhotographerService
+import com.hm.picplz.data.provider.DummyPhotographerProvider
 import javax.inject.Inject
 
 interface PhotographerSource {
@@ -9,8 +9,9 @@ interface PhotographerSource {
 }
 
 class PhotographerSourceImpl @Inject constructor(
-    private val service: PhotographerService
+//    private val photographerApi: PhotographerApi,
+    private val photographerApi: DummyPhotographerProvider
 ) : PhotographerSource {
     override suspend fun getPhotographers(): Result<PhotographerListResponse> =
-        runCatching { service.getPhotographers() }
+        runCatching { photographerApi.getPhotographers() }
 }
